+++ lib/lib_cleanex.ex
@@ -1,5 +1,5 @@
 defmodule Cleanex do
   @doc "Clean a term by invoking the Cleanable protocol"
-  @spec clean(term) :: term
-  def clean(term), do: __MODULE__.Cleanable.clean(term)
+  @spec clean(term, keyword) :: term
+  def clean(term, opts \\ []), do: __MODULE__.Cleanable.clean(term, opts)
 end