+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.4.0 (September 24, 2019)
+======================
+* Add plug option `[struct: true/false]`. (Default is `true`)
+  When set to false, inner whitespace won't be touched.
+
 0.3.0 (August 3, 2018)
 ======================